Refactor helper_functions.py into domain-specific utility modules
- Move LLM output processing functions to llm_utils.py (remove_reasoning_from_output, clean_json_tags, clean_markdown_tags, safe_json_loads)
- Move HTML extraction to css_utils.py (extract_html_from_content)
- Create new config_utils.py for configuration management (load_pipeline_config, check_required_env_vars)
- Update all imports across the codebase
- Delete the now-empty helper_functions.py

This improves code organization by grouping related functionality together.
