+++ deep_research/steps/cross_viewpoint_step.py
@@ -4,10 +4,7 @@ import time
 from typing import Annotated, List
 
 from materializers.analysis_data_materializer import AnalysisDataMaterializer
-from utils.helper_functions import (
-    safe_json_loads,
-)
-from utils.llm_utils import run_llm_completion
+from utils.llm_utils import run_llm_completion, safe_json_loads
 from utils.pydantic_models import (
     AnalysisData,
     Prompt,
