+++ deep_research/utils/llm_utils.py
@@ -1,15 +1,11 @@
 import contextlib
 import json
 import logging
+from json.decoder import JSONDecodeError
 from typing import Any, Dict, List, Optional
 
 import litellm
 from litellm import completion
-from utils.helper_functions import (
-    clean_json_tags,
-    remove_reasoning_from_output,
-    safe_json_loads,
-)
 from utils.prompts import SYNTHESIS_PROMPT
 from zenml import get_step_context
 
@@ -20,6 +16,81 @@ logger = logging.getLogger(__name__)
 # ALL model names require a provider prefix (e.g., "sambanova/", "openai/", "anthropic/")
 
 litellm.callbacks = ["langfuse"]
+
+
+def remove_reasoning_from_output(output: str) -> str:
+    """Remove the reasoning portion from LLM output.
+
+    Args:
+        output: Raw output from LLM that may contain reasoning
+
+    Returns:
+        Cleaned output without the reasoning section
+    """
+    if not output:
+        return ""
+
+    if "</think>" in output:
+        return output.split("</think>")[-1].strip()
+    return output.strip()
+
+
+def clean_json_tags(text: str) -> str:
+    """Clean JSON markdown tags from text.
+
+    Args:
+        text: Text with potential JSON markdown tags
+
+    Returns:
+        Cleaned text without JSON markdown tags
+    """
+    if not text:
+        return ""
+
+    cleaned = text.replace("```json\n", "").replace("\n```", "")
+    cleaned = cleaned.replace("```json", "").replace("```", "")
+    return cleaned
+
+
+def clean_markdown_tags(text: str) -> str:
+    """Clean Markdown tags from text.
+
+    Args:
+        text: Text with potential markdown tags
+
+    Returns:
+        Cleaned text without markdown tags
+    """
+    if not text:
+        return ""
+
+    cleaned = text.replace("```markdown\n", "").replace("\n```", "")
+    cleaned = cleaned.replace("```markdown", "").replace("```", "")
+    return cleaned
+
+
+def safe_json_loads(json_str: Optional[str]) -> Dict[str, Any]:
+    """Safely parse JSON string.
+
+    Args:
+        json_str: JSON string to parse, can be None.
+
+    Returns:
+        Dict[str, Any]: Parsed JSON as dictionary or empty dict if parsing fails or input is None.
+    """
+    if json_str is None:
+        # Optionally, log a warning here if None input is unexpected for certain call sites
+        # logger.warning("safe_json_loads received None input.")
+        return {}
+    try:
+        return json.loads(json_str)
+    except (
+        JSONDecodeError,
+        TypeError,
+    ):  # Catch TypeError if json_str is not a valid type for json.loads
+        # Optionally, log the error and the problematic string (or its beginning)
+        # logger.warning(f"Failed to decode JSON string: '{str(json_str)[:200]}...'", exc_info=True)
+        return {}
 
 
 def run_llm_completion(
